12.26 完善精度

+++ Cal_ViewModel/OperatorModel/ComplexOperation.cs
@@ -26,7 +26,7 @@ namespace Cal_ViewModel.OperatorModel
         }
         public string Squ(string num1)
         {
-            return (Convert.ToDecimal(num1) * Convert.ToDecimal(num1)).ToString();
+            return (Convert.ToDouble(num1) * Convert.ToDouble(num1)).ToString();
         }
         public string Cube(string num1)
         {
